requirements.txt
streamlit
pandas
plotly_express
plotly.graph_objs
